requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
aniso8601==9.0.1
async-timeout==4.0.2
attrs==22.2.0
black==23.3.0
cachelib==0.10.2
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
discord-webhook==1.1.0
dnspython==2.3.0
Flask==2.2.3
Flask-Discord==0.1.69
Flask-Minify==0.41
Flask-PyMongo==2.3.0
Flask-RESTful==0.3.9
Flask-Session==0.4.0
frozenlist==1.3.3
htmlmin==0.1.12
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
jsmin==3.0.1
lesscpy==0.15.1
MarkupSafe==2.1.2
multidict==6.0.4
mypy-extensions==1.0.0
oauthlib==3.2.2
openai==0.27.4
packaging==23.1
pathspec==0.11.1
platformdirs==3.2.0
ply==3.11
PyJWT==2.6.0
pymongo==4.3.3
python-dotenv==1.0.0
pytz==2023.3
rcssmin==1.1.1
requests==2.28.2
requests-oauthlib==1.3.1
six==1.16.0
tomli==2.0.1
tqdm==4.65.0
urllib3==1.26.15
Werkzeug==2.2.3
xxhash==3.2.0
yarl==1.8.2
